Refactor: split up allow-call-syntax handling to parse-lhs-expr.
`js2-parse-member-expr` used to handle all 3 productions `NewExpression`,
`MemberExpression`, and `CallExpression`, the last being toggled by
`&optional allow-call-syntax`.  This commit splits up `CallExpression` from
it and rename it to `js2-parse-member-or-new-expr`, so that future changes
in the standard to these productions are easier to work with; two new
functions `js2-parse-lhs-expr` and `js2-parse-lhs-expr-tail` are added as
an API similar to the old, and the renamed `js2-parse-member-or-new-expr`
no longer parses a `CallExpression`.